requirements.txt
torch
tqdm
PyYAML
pygame

